smartclass-frontend/src/Frontend/requirements.txt
click==8.0.1
docutils==0.14
Flask==2.0.1
importlib-metadata==4.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
lockfile==0.11.0
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
protobuf==3.17.0
pystache==0.5.4
python-daemon==2.2.3
simplejson==3.2.0
six==1.16.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
